# p2p (#4119): treat slow integration tests