--- a/p2p/dial_integration_test.go
+++ b/p2p/dial_integration_test.go
@@ -1,0 +1,10 @@
+//go:build integration
+// +build integration
+
+package p2p
+
+import "time"
+
+func init() {
+	dialTestDialerUnexpectedDialTimeout = 150 * time.Millisecond
+}